AG-15406 - Add flash on update to options
